requirements.txt
PyQt6
PyMuPDF
numpy
pillow
wordcloud
